+++ internal/tui/views/constants.go
@@ -9,6 +9,7 @@ const (
 	viewPaneSwapPicker
 	viewConfirmKill
 	viewConfirmCloseProject
+	viewConfirmCloseAllProjects
 	viewConfirmClosePane
 	viewConfirmRestart
 	viewHelp